app/requirements.txt
jwcrypto
Flask
line-bot-sdk
boto3
